KIALI-635: Fix namespace chart lost
After the graph is loaded, when the user clicks the background,
properties are pushed to the component only once. Because of this, the
side panel was not loading the required data.

Changing some code to ensure that "final" properties are always pushed
when the side panels are mounted. This allows to load chart data once
the component is mounted.