KIALI-1283 Correctly link to service details page
Now, nodes in the graph can be of several types. Nodes will be backing
services, so the old links in the headers of the side panel are no
longer valid.  Links to the backing services are now listed in the body
of the side panel (one node can be backing several services).

This commit is also: removing the link to the detailed charts for the
"edge" version of the side panel, since a link cannot be resolved
reliably; fixing titles-headings of the side panels with the right
legends (app, workload, servce).