KIALI-1253 Fix missing warning icon in VS/DR tab
Cheks that issue warnings do not necessarily mark a
VirtualService/DestinationRule as invalid. So, in order to show the icon
in tab headers, presence of any failed check is verified.

Right now, the backend is only returning checks that have 'failed'
either with low or high severity. Checks that pass are not informed. This
makes safe to assume that if the backend has returned something, then at
least, a warning icon can be shown.