KIALI-1156 Fix behavior of message center; add system errors panel.
Per some discussion, the notification bell and the "open issues" message
aren't following the pattern suggested by UX. This fixes the behavior of
the messsage center, and the notifications in the main header of Kiali
page.

* Regular kiali errors and warnings (mainly, network errors) will now
show the standard bell with a badge when there are messages in the
message center. If there are errors or warnings, the badge will be red;
other types of messages will show the blue badge (we don't have any of
those kinds as of now). This is the standard PF behavior of the
notification bell and notification drawer.
* The warning "open issues" won't replace the notifications bell. For
now, the only issue that raise this kind of warning to show up is the
unsupported Istio version warning (reported by the backend).
* An "Open issues" panel is added to the message center. This panel is
visible only when the Istio version warning is reported by the backend.
Messages in this new panel can be marked as read, but cannot be cleared.
Open issues panel is open by default when clicking in the "open
issues" message. "Notifications" panel is open by default when clicking
the bell
* Marking open issues as read will clear the notifications count, but
won't clear the "open issues" message. This is to remind the user that
there is something wrong.
* Backend status is now fetched on browser refresh, if user is already
logged in. This is to fetch and raise the Istio version warning, if
needed. This ensures that the user is aware of the problematic
condition.