KIALI-1380 Add links to the website and github organization
Links are added to the 'about' dialog.

Links, icons and text are configurable in the config.ts file. The 'about' section config.ts file can be absent completely, or have only the webpage or project sections. The about box will hide and show what is available in the config file.