KIALI-1429 Fix channel read race condition
In a "select block" of channels, Go does not guarantee any priority of evaluation. This was leading to a race condition when reading between the "done" channel and the data channels: the "count" variable may be incremented even before respective data was read. This was leading to early termination of the for loop, causing not all data being collected.

The extreme case was if Go evaluates first the "done" case  on all iterations. Under this circumstance, no data will be read and getServiceHealth would return EnvoyServiceHealth{} with zeroes.

The channel read-race condition was also leading to sporadic test failures in CI.

To fix the issue, the "done" channel is removed by ensuring that data and error channels always return a value.