KIALI-1714 Limit the lower bound of Prometheus queries (#560)
Lower bound is set only if needed, in order to not query time ranges that don't fall into the time of life of the namespace whose data is being queried.

Both, prometheus query ranges and rate intervals are subject to be adjusted. For instant queries, the rate interval can be adjusted to not let fetch data before the creation date of the namespace.

For range queries, both the start time of the query and the rate interval could be adjusted. The rate interval is only changed if, from the "end date" of the range, the query would try to fetch data before the creation date of the namespace. The start date will be adjusted if, in combination with the rate interval, the query would fetch data before the creation date of the namespace.