KIALI-1670 add support for groupBy options (#820)
* KIALI-1670 add support for expanded app box (groupBy=app)
- as per feedback, we now apply app grouping automatically to
  app and versionedApp graphs, and no grouping to service and
  workload graphs. (the alternative is adding a 'groupByApp'
  option to the settings dropdown, but this was deemed potentially
  confusing and we'll wait for feedback)
- fix to redux init on injectServiceNodes change
- fix (supplied by Josejulio) to GroupCompoundLayout to protect against loopback edges
- minor: some alphabetical ordering changes here as well

* Ensure graph is reset whenever we change namespace.  It is not ideal to
have to dispatch two actions whenever we want to set the activeNamespace. Ideally
all of the relevant state should be updated at one time.  But I'm not sure
what the best approach is, and I think we need another round of state handling
revision anyway, best for a separate PR.