KIALI-2188 Fix root path redirection loop
Because of the `StrictSlash(true)` on the root URL, the muxer was redirecting from `$webRoot/` -> to -> `$webRoot` (without trailing slash). However, there is a manual redirection in the opposite way when $webRoot is not '/' (custom context case). This was causing a redirection loop making impossible to access the UI using the root Kiali URL when custom context is configured.

 The loop is removed by turning on "StrictSlash" after the manual redirection is configured.