KIALI-2352 Update prom rate query param calculation given change in scrape interval Istio 1.1 has updated the scrape interval for relevant metrics from 5s to 15s, presumably to reduce metric collection intensity.  This affects the rateInterval/step param calculation we do in the UI, resulting in "single-dot" sparklines for "Last Minute" durations, and affects other calls as well.
The change increases the minimum step to include 2 data points (making rate()
for short durations basically the same as irate()).  This provides us with at
least a minimal sparkline graph of 3 points.

Additionally, do a little refactoring for this compute method and set things
up for a possible follow-on that adds scrap interval to the server config.