KIALI-2479 Block console rendering until all configs are retrieved (#1043)
* KIALI-2426 Get serverConfig from static env.js

So this is removed from redux (lot of changes) and used instead as a standard constant.
This constant is written in 'window' object, in env.js, by the server. File is imported in index.html as static js.

Valid durations are now computed statically once for all.

* Validate Duration at URL-parsing, not at render

While adding functions in HistoryManager to validate URL, taking that opportunity to do some refactoring here:
- Graph components were calling functions from ListPageHelper, which it isn't supposed to do, so moving these functions to HistoryManager
- Rename enum URLParams => URLParam
- More homogene usage of "undefined" instead of "null" in getting URL params

* KIALI-2479 Block console rendering until all configs are retrieved

This is adding an AuthenticationController component. This component
becomes responsible on determining the moment when it's OK to render the
console.

Code in charge of retrieving server configs is integrated into the
AuthenticationController component.

Configs from /api/grafana and /api/status are not required for Kiali to start. This tolerates errors from these endpoints and proceeds to rendering of the main console.