Graph generation has two steps.  Step one is to generate the TrafficMap by querying the provided telemetry. For example, querying Prometheus for Istio Telemetry.  This will differ based on the mesh/telemetry provider.  Step two is transforming the TrafficMap into a vendor-specific format.  For example, the Cytoscape JSON.
This PR relates to step one.  We need to clearly define the TrafficMap to more
easily allow for multiple generators.  The new Meta.go is a baby step in that
direction.  It defines the non-protocol keys that can be populated in the Node and Edge
metadata.  Protocol.go is responsible for defining the protocol-specific keys.

The new Meta.go also ensures that we can use constants as opposed to string literals
when populating the Metadata, which is much more safe.