KIALI-2858 Restructure backend graph code for better separation of core and vendor code (#1068)
Producing a graph has two main phases:

Telemetry processing
Generation
Telemetry processing takes a vendor's telemetry and produces a TrafficMap. The TrafficMap is a Kiali construct that describes all of the traffic (for the time period): nodes, edges, rates, badging, etc.. (current vendors: Istio)

Generation takes the TrafficMap and creates a vendor-consumable configuration that can be rendered by the vendor's tool. (current vendors: Cytoscape)

This PR mostly relates to step one. There are a few goals:

Clearly define the TrafficMap such that it can:
- Be produced by different telemetry vendors.
- Be consumed by different generation vendors.

Also:
- Separate core code from vendor code
- Clean up some code and simplify packages

To help define the TrafficMap this PR introduces meta.go. This new core file defines the non-protocol keys that can be populated in the Node and Edge metadata. It also ensures that we can use constants as opposed to string literals when populating the Metadata, which is much more safe. Protocol.go (which existed prior to the PR) is another core file and is responsible for defining the protocol-specific keys.

It also introduces 2 new interfaces: TelemetryVendor and ConfigVendor, for implementations of the two required APIs for vendors. The Istio code now implements TelemetryVendor and the cytoscape code now implements ConfigVendor. These complement the existing Appender interface.

The updated code is currently structure like:

/handlers
    graph.go        # rest api handler
/graph
    ...                    # core files
    /api                # entry point api (called from rest api handler)
    /telemetry
        /istio          # istio telemetry impl
    /config
        /cytoscape # cytoscape config impl