KIALI-2930 Show details of error in initialization screen
At first load of Kiali, a request is done to retrieve some settings. If this request fails, an error message is shown. Sometimes, this error message is not enough to troubleshoot the issue and the developer console is needed to see what the back-end is returning.

To improve this experience, adding a "Show details" pane to display the raw response sent by the back-end if first-load initialization fails.