KIALI-2529 Fix type erasure on selectors
This is extracted from #1154, using the `unknown` type added on
Typescript 3.0 (we're using 3.4.5). It fixes a problem that happens
after the change from `react-scripts-ts` to `react-scripts` and breaks
the build there.

There should not be any breakage, as the generated code is the same. The
only thing this is doing is helping the compiler to understand what
types are being returned.