KIALI-2999 Fix issues with ServiceEntry identification (#1137)
* KIALI-2999 Fix issues with ServiceEntry identification
Given a telemetry destination_servicename we attempt to determine if that
destination is actually a serviceEntry.  We do it by matching the service
name to the hosts defined for the accessible serviceEntrys.  I'm pretty sure
this PR does not cover all of the cases but it is an improvment over the
exact string match previously required.  This PR adds:
- wildcard handling for external service entries
- prefix handling for internal service entries with host set to the fully qualified service

This also fixes two bugs:
- we were failing to cache serviceEntry info in the appender globalInfo
- we called for the serviceDefinition of a serviceEntry, which does not exist