KIALI-2859 KIALI-2159 Replace gometalinter with golangci-lint (#1164)
* KIALI-2859 Replace gometalinter with golangci-lint

https://issues.jboss.org/browse/KIALI-2859

Also, fixing the newly reported issues with the additional linter introduced by golangci-lint.

From the default linters set, only "errcheck" is disabled because it generates too many errors to fix under the scope of this PR/JIRA.

* KIALI-2159 Move to a supported golang version in CI

https://issues.jboss.org/browse/KIALI-2159
Since Go 1.9.4 is unsupported for golangci-lint. Also, 1.9.x is currently unsupported. Thus, moving to 1.12.1 which is used in the build pipeline.