KIALI-3150 Fix blank screen
https://issues.jboss.org/browse/KIALI-3150

There is some issue with details pages. Lazy loading all them produces a blank screen bug in the containerized build (dev-mode `yarn start` works fine). It's enough if one of the detail pages is not lazy loaded to have the app working fine.

For now, leaving AppDetailsPage without lazy load to fix the issue. I'll try to find what's the root cause, but doing this to not delay the fix.