Gateway validations: tighten up duplicate host matching regex. (#1284)
* Gateway validations: tighten up duplicate host matching regex.

This is to address an issue where a perfectly legitimate `Gateway`
with hostnames `example.com` and `foo.example.com` would get flagged
with a duplicate-host warning.  This was as a result of an unanchored
regular expression.  This PR introduces a test case to illustrate it,
as well as a fix.

Closes #1272.

* business/checkers/gateways/multi_match_checker.go (findMatch):
  Anchor the hostname with `^$` when it is used as a regular
  expression.  This avoids incorrect substring matching behaviour.
* business/checkers/gateways/multi_match_checker_test.go (TestNoMatchOnSubdomainHost):
  Introduce a regression test of the above.

* Add another wildcard/subdomain combination test.

* business/checkers/gateways/multi_match_checker_test.go (TestAnotherSubdomainWildcardCombination):
  Add test case.