KIALI-3170 Fix operator deploy script to fetch last release
Deploy script was taking the last release from a chronological point of view, which may not be the release we want users to install since it may not be the latest release using semantic versioning logic.

This, hopefully, fixes the deploy script to take the latest release using semver logic.