Add NPM registry parameter
Adds NPM_CONFIG_REGISTRY parameter to the sample job.

Unfortunately, since different envs may need to use different package registries it's no longer possible to hardcode the parameters to make the Pipeline to "auto configure" itself if creating manually (harcoding configs replaces any customized configs of the Pipeline on every run)

This means, that the Pipeline will need to be configured manually if the config.xml file cannot be placed in the Jenkins instance.