Do not do prefix matching for multi-cluster ServiceEntry
It was assumed that MESH_INTERNAL ServiceEntries refer to services in the local cluster. But when Istio is setup in a multi-cluster scenario, MESH_INTERNAL ServiceEntries are also used to route traffic to services in remote clusters.

ServiceEntries that route traffic to remote clusters must have a hostnames configured of the form remote_svc_name.remote_ns_name.global. I.e. always have three entries and last entry is the static string "global". This helps to clearly differentiate ServiceEntries are referring to local services and the ones that are routing traffic to remote clusters. This changes are implementing this logic.

Fixes #1364, #1365, #1366, #1367