Re-label telemetry destination_service={*.global} to hostname only
In a multi-cluster scenario, in the cluster that is the destination of
the traffic, Istio will record telemetry with
destination_service="svc.namespace.global". This was generating a graph
with dummy service nodes labeled with "svc.namespace.global". These
represent the same "svc" kubernetes services without any suffix.

This changes perform some renaming to generate a traffic map that will
lead to proper unified service nodes.

Resolves #1481.