Capture and report request.host information when it is configured
Users have no way to easily see the requested host for their requests. This
can be very useful for failed requests, especially external requests, but is
also has informational value in other cases.

The request.host info must (currently 1.3) be configured manually for Istio, but
if done this PR will capture the information and report it back as part of
the "responses" data, along with rates and response_flags.

https://github.com/kiali/kiali/issues/1405