Capture and report request.host information when it is configured (#1687)
Capture and report destination_service information, which starting in Istio 1.3.1, contains
request.host information for PassthroughCluster and BlackHoleCluster, which can be very valuable for understanding egress traffic.  Users previously had no easy way to see
the requested host for this traffic.  For most traffic this will show the destination.service.host,
which is still useful.  The information is reported back analogously to the response_flags info,
broken down by response_code, as part of the traffic data.

https://github.com/kiali/kiali/issues/1405