Implementing edge aggregation for nodes "talking" to a ServiceEntry (#1756)
One Istio ServiceEntry can have several hostnames in its definition, leading to an initial traffic map with one service node for each hostname in the SE. The service_entry appender was correctly reducing the service nodes into one ServiceEntry node, but edges were only being re-bounded leading to the possibility of multiple edges between the same source node and destination SE node.

This is adding a protocol-aware edge aggregation to remove redundant edges. It's also adding the protocol-awareness for the SE-to-Egress case.

Fixes #1752.