PF4 Help Tours (#1515)
Migrate to a PF4 Popover approach.  Because Popover does not support HTML targets, but rather needs to be wrapped around children elements, the approach had to change somewhat.  A Tour is still declared in it's own file, with an ordered set of TourStops, but additionally each stop on the tour must be instrumented with a surrounding TourStopContainer.
- replace the DIY state management with redux
- no PF4 tour design, so maintain PF3 design with tour stop number in title
- known issues
  - there seems to be a bug in Popover such that outside clicks sometimes hide the element, I
    added a workaround that can cause a slight "flash".  I will follow up with PF.
  - the "disappearing masthead" issue (https://github.com/kiali/kiali/issues/1834) happens if the
    tour is active and graph Display dropdown is clicked.  This is likely a PF4 bug and is being
    researched separately.  An F5 is required to redraw.