Fix BrowsersList warning and other package.json updates (#1516)
* Rapidoc upgrade

Bumping to latest 5 major version of rapidoc to remove manual resolution.

* Remove caniuse-lite resolution

History in git doesn't how the reason why this resolution was included. Without it, yarn resolves to a more recent version.

* Bump version of react-scripts dependency from 3.1.1 to 3.1.2

* Bump version of react-scripts dependency from 3.1.2 to 3.2.0

* Run yarn upgrade react-scripts

This upgrades versions of react-scripts dependency tree.

* Fix resolutions and dependencies

* Remove @typescript-eslint/* manual resolutions, which were added to fix a typescript version warning, which is no longer present, given that current version of react-scripts have upgraded these sub-dependencies.
* Add table-resolver@3.3.0. This is sub-dependency of patternfly-react, for building tables. Since it's not being referred as a direct dependency, it became unavailable after `yarn upgrade`. So, adding it to fix build error.

* Remove handlebars from resolutions

Sub-dependency of react-scripts, which has already been bumped. So, the forced resolution is no longer required.

* Remove js-yaml from resolutions

No need for it because it's already a dependency.

* Fix yarn.lock after rebase

* Remove table-resolver dependency

After rebasing, GraphHelpFind is on PF4. The table-resolver dependency is no longer needed.