Improve handling when graph summary target element changes (#1545)
When transitioning from one selected graph element to another of the same type (e.g. edge1 to edge2), we rendered one time using stale metric data.  This was invalid in a normal case, and in the worst case could cause a crash.  This PR ensures that on a prop change for the target element we immediately move to loading state, and initialize the metric state.  It also ensures a more complete init of metric state when applicable.  It applies the same logic to edge, node and
group (i.e. app box) summary panels.

https://github.com/kiali/kiali/issues/1933