Fix templates
Lately, looks like no labels are being added to new issues when using the templates. But issue templates have labels that no longer exist because the emoji was removed some time ago.

Probably, this is the reason behind no labels being added when using the templates. This may fix the problem and we may, again, see the labels automatically added.

Related #1969.