Add console links to OS 4.3
This commit adds kiali links to the OpenShift console, one in the
Application Launcher for the dashboard, and a contextual link on the
project details page, which opens directly to the graph on the
namespace.

There are some current limitations on the implementation:

* There is no way to set an icon to the app launcher item.
* There is no way to have a placeholder on the link to set the correct
  namespace, so I needed to create one `ConsoleLink` for each namespace
  in the cluster.
* I found a bug where logging in with a link for an internal page
  redirects the user to the dashboard, which breaks behaviour if the user
  is not signed in when it clicks on the contextual link.

The UI implementation for this (adding the ApplicationLauncher to
Kiali-UI) still depends on external input from the Console team.