Fix error: Kiali cache doesn't support sidecars
Need to change the logic in istio_validations, as istio objects can be part or not of cache, we need to fetch types, in cache or not, one by one and not all at the same time.

In the end, some logic in kubernetes package is moved to business package.

Fixes https://github.com/kiali/kiali/issues/2488