Move "Show on graph" link down to the mini-graph card (#1676)
* Move "Show on graph" link down to the mini-graph card

The link was moved under a kebab menu.

Related to kiali/kiali#1960.

* Disable tap and double-tap events on mini-graph

Events are disabled because when in mini-graph mode, only pan and zoom should be available and no other events should happen.

* Create MiniGraphCard component to reduce code duplication

The mini-graph is present in app, workload and service detail pages. The mini-graph related code is very similar on all these pages. This is extracting the mini-graph to a component to reduce code duplication.

Also, in CytoscapeGraph component, the `activeNamespaces` and `graphType` properties were removed because these data is present in the datasource.

* Fix tests

* Remove unused state from AppDescription.tsx and ServiceInfoDescription.tsx

* Fix UXD feedback:

* Use "Show full graph" instead of "View full graph" in mini-graph's kebab menu
* Turn off traffic animation for mini-graph
* Shorten to "Back to full graph" on the back button of the drilled-down graph
* Use "Show" instead of "Go to" in the cards of the ovewview page