Add compatibility for Telemetry V1 on Istio 1.5
IstioSupportsCanonical function is updated to read Istio's config map and check a flag that dictates if telemetry through mixer is enabled (aka Telemetry V1). If flag is true, TelemetryV2 is active.

IstioSupportsCanonical is using Kiali's Service Account always rather than the account associated with the logged in user. This assumes that Kiali's ServiceAccount has privileges to read Istio's namespace, in contrast to reduced privileges when using authentication methods that involve the RBAC system. It should be so-so safe to do this, because Kiali ServiceAccount tends to always have privileges in Istio system namespace (at least, this is the default when first installing Kiali).

Fixes kiali/kiali#2669