Graph Hide can crash on graph refresh (#1742)
Graph hide (with compressOnHide) keeps the hidden elements as "removed" elements in cy. The "removed" elements can later be restored when the hide-expression is changed or deleted. It's only safe to remove/restore to the same set of graph elements, otherwise there can be ID clashes or other consistency problems (read: crash). This PR fixes issues where the graph-hide logic could try to remove or restore elements when the graph had changed or was in the process of being updated. There are several changes due to additional issues found while testing earlier revisions of this PR. Ultimately we needed to make changes in how we utilized GraphDataSource in conjunction with react and redux

In the end, the graphHide feature is delicate because it is processed as part of toolbar updates but extremely sensitive to changes in the cytoscape graph. This PR maintains that approach and tries to make it all work. But, if I were to start again I may have tried to move the actual "handleHide/HandleFind" logic into the core graph processing (CytoscapeGraph) and limit the toolbar to only manage the expressions, storing that info in redux and then letting the graph code use it. Most if not all of the problems result in the graph updating in one place and the hide logic updating and manipulating the graph in another.

https://github.com/kiali/kiali/issues/2656