Fix an issue with zoom/pan/fit. (#1754)
This fixes at least one issue with zoom/pan/fit. When using graph-hide we only want to re-establish the zoom/pan settings if we believe the graph is unchanged (basically the same removed elements are in play as when we previously executed the hide logic).  The determination is not perfect but should work fairly well.
- also made a couple of minor tweaks to "safeFit()" on the off-chance it  could help.

Related to https://github.com/kiali/kiali/issues/2514