Remove RBAC limitation from the token authentication
This is for plain Kubernetes environments, where the Projects API is not available. There was a RBAC limitation where Kiali was requiring read access to all Kiali accessible namespaces in order for users to be able to login.

This tries to circumvent the RBAC limiation by using the Kiali Service Account. Certainly, this is assuming that the Kiali ServiceAccount is present and has the `get namespaces` privilege. We are already assuming the presence of Kiali ServiceAccount with other sets of permissions for the following features:

* Kiali's cache
* Discovery of prometheus
* Discovery of Grafana
* Check what version of Istio's Telemetry is enabled

So, the assumption of the presence of the Kiali ServiceAccount and it's usage may not be too bad. Installation through Kiali Operator is currently creating the SA, at least with read-only privileges. Anyway, if the Kiali ServiceAccount is missing, or not properly mounted, or it doesn't have the assumed privileges, this changes should still fail gracefully.

Fixes kiali/kiali#2728