Navigation to other detail pages using mini-graph (#1826)
* Navigation to other detail pages using mini-graph

Double clicks on mini-graph card is enabled. When double-clicking, the mini-graph will redirect the user to the detail page of the selected node. No redirection is done if the detail page of chosen node is the one already being displayed. Also, nodes with no detail page (like service entries and inaccessible nodes) are ignored.

A small refactor had to be done to the CytoscapeGraph component. The double-tap event handlers were pulled up to the GraphPage component in order to be able to perform a different action depending on the page hosting the CytoscapeGraph component.

Fixes kiali/kiali#2790.

* Switch to single-click for navigation on mini-graph

Preferring single-click vs double-click.

Not doing the same refactoring like in double-tap because it would involve more changes. But keeping the double-tap refactor just in case and because it seems good to have.

Related to kiali/kiali#2790.