Fix test to determine OpenId/OAuth preparation
OAuthLogin class is being used for both openshift and openid login strategies. OpenShift passes always the `access_token` parameter. OpenId providers *may* pass the access_token parameter, but id_token is warranted to be present (because of the asked response_type).

This is fixing a blocked login if the OpenId provider is not passing the access_token.

Fixes kiali/kiali#2925