Fix Snyk
Running Snyk within CI, because Snyk's GitHub integration does not
support go modules.

If PRs are coming from forks, Snyk will run when PRs are merged to
master. If PRs is within kiali/kiali repo, then Snyk can run on the PR.
This is because of how Travis secrets work.