Better handle egress destination telemetry (#3004)
Istio Egress destination telemetry is lacking and may not be easily fixed on their side. So, add some special case code to help things out for the Kiali graph.  Note that the bad telem is still in Prometheus, so direct metric query won't match up completely unless we want to patch that as well.

https://github.com/kiali/kiali/issues/2999