Add support for cluster API proxies on OpenId strategy
Since some cloud providers won't support configuring OpenID integration in the managed Kubernetes, users may want to use workarounds. One of them is to place a proxy in front of the Kubernetes API handing the OpenID auth (an example is kube-oidc-proxy).

This is adding support for such API proxies.

Fixes kiali/kiali#3042