Initial implementation of OpenId's auth code flow (#3189)
* Initial implementation of OpenId's auth code flow

This initial implementation of the "authentication code" flow for OpenID authentication let's Kiali negotiate authentication with an extenal IdP provider. Kiali is acting as an "unauthenticated client" (i.e. no client secret is used).

This is using AES-GCM encryption to encode the cookie which prevents revealing the OpenId id_token to the network and to the browser.

Related to #3088

* Refactor: cleanup and prepare to re-use some code

Related #3088

* Mazz feedback: Unique error messages

Co-authored-by: John Mazzitelli <mazz@redhat.com>

* Feedback: Jay & Mazz

* Feedback: Fix message

Co-authored-by: John Mazzitelli <mazz@redhat.com>