Support for acting as authenticated OIDC client (authorization code flow) (#3215)
* Support for acting as authenticated OIDC client

When OIDC client should authenticate, we need the client-secret provided/configured by/in the OIDC server. For this, we re-use our old/reserved "kiali" secret that was used in the past for the already removed login authentication strategy avoiding another set of changes in the operator.

This also:
* Fixes logout
* Fixes broken OIDC "implicit flow" (nonce cookie deleted early)

Related #3088

* Unify oidc implicit flow code (deduplicate)

This commit reuses the functions added to openid_auth.go created for the authorization code flow. Most of that code is shared. It's just the order of execution that changes. So, this is a small rework of performOpenIdAuthentication function in authentication.go file to remove code duplication.