add 'healthy' operand to graph find/hind / ensure health for graph nodes is part of the overall graph fetch (#1957)
* Support 'Healthy' operand in Graph Find/Hide

To be able to do this we need the node health as part of the graph fetch.
This is because find/hide is applied immediately after fetch.  Health is
pulled async once we have the graph nodes.  Prior to this it was done
independently of the graph fetch, now we ensure the health info is
attached prior to emitting 'fetchSuccess'. It may slow down graph fetch
but we gain consistency.  Going forward we will need to ensure graph fetch
is not overly impacted in a negative way.

Health fetch migrates from CytoscapeGraph to GraphDataSource. It can be
optionally fetched as not every user of GraphDataSource needs the health
information.