Add multi-thread safety to OpenId functions
Multi-thread safety is added to:
* GetJwkFromKeySet (the new one in #3084)
* GetOpenIdMetadata (an existent one)

It's unlikely that an error will happen without this, but it's better to be safe. The main variables thay may lead to race condition issues are `cachedOpenIdMetadata` and `cachedOpenIdKeySet`. If multiple threads want to read these variables and find that no data is cached, then multiple requests will be dispatched to fetch the missing data and also all threads will want to store the fetched data.

These changes are to ensure that only one request will be dispatched to fetch the data even if multiple threads find that there is no cached data. The same fetched data will be returned/shared across all threads that found no cached data and are wanting to retrieve it from the OpenId Server.

Related to kiali/kiali#3084