Fix the path of the cookie for OpenShift authentication strategy
The missing path directive for OpenShift authentication is creating a cookie targeting https://{fqdn}/{webroot}/api (notice the "api" sub-path).

When logging out, cookies are cleared using the Kiali root path (i.e. https://{fqdn}/{webroot}). This causes a mismatch in the browser and the session cookie is not cleared for OpenShift strategy. As a consequence, Kiali ends in a bad state where it's not possible to login again (workaround is to manually clear browser cookies).

This fixes the creation of the session cookie for OpenShift by specifying the root path. Also, when deleting the session cookie, in addition to setting an expiry time in the past, the MaxAge attribute is also passed with negative value.

Fixes kiali/kiali#3313