[hack] get the install istio script to work with 1.8. (#3393)
This rips out some old stuff for 1.4 which we don't care about anymore.
This also changes the --addons so it no longer always adds prometheus - someone asked if it was possible to not install Prometheus and there was no way to do so before. Now there is.
component.telemetry is mixer-related - mixer is gone in 1.8, so we can only use components.telemetry on 1.7 and earlier.
global.controlPlaneSecurityEnabled seems to have gone away.