Function to resolve what remote clusters are visible
This resolves remote clusters by fetching the list of "remote secrets" that are present in the istio namespace.

Apparently, remote secrets are labeled with istio/multiCluster=true (at least, this is the label that presets the `istioctl x create-remote-secret` command). So, this is fetching secrets with that label and assuming they contain the kubeconfig files required by istio to contact the remote API Server and enable discovery.

Some better error handling is pending.

Related #3409