Resolve the "Home cluster" in an multi-cluster env
"Home cluster" means: the cluster where the "adjacent" control plane is running.

"adjacent control plane" means the control plane that Kiali is observing/managing.

This is a draft about how to resolve the "home cluster". In general, the clustering.go file is assuming that Istio is configured with multi-cluster and makes assumptions that all configs are there. If all configs are set, this should work OK.

Related #3409