Resolve the network of discovered clusters
In multi-cluster enabled meshes, each cluster is assigned to a network. The Control Plane requires a logical network ID to be assigned to each network in the mesh.

This changes find what's the logical network ID configured on each discovered cluster and is reported in the response JSON.

Related kiali/kiali#3448