Add proxy to OpenId metadata http requests
This PR adds proxy support to OpenId's requests for metadata for the
issuer. This fixes the problem where Kiali requires a proxy to connect
to the outside world. This is not normally a problem, as the resources we
connect using Kiali are usually inside the cluster.

OpenID is the exception, as we need to connect to the issuer to get
information to move the user forward. This PR adds proxy support to this
specific request.

** Possible Questions

1. Are there any other places that should also respect the proxy?
  Not that I'm aware of, I did some lookaround on this but could not
  find other place where it is immediatelly obvious to connect to places
  outside the cluster.

2. How to test this?
  1. Install Squid on your machine, or other HTTP proxy.
  2. Install kiali with the correct configuration options (`http_proxy`
  and `https_proxy` on the openid section of the CR, and OpenId configured).
  3. Connect/authenticate on Kiali using OpenID
  4. Check squid log for the request.