Rollback a few changes to NewClientFromConfig
These changes are, apparently, unneeded. Also, it's a error prone, because there were two tokens involved when creating the client. There was one in the `config` parameter, and another one in `authInfo` parameter. Only the token in `config` was being used. The `authInfo` is used only to pass impersonation data, but this data shouldn't always be honored because it's only valid when using the `header` strategy.

So, I think it's better to rollback these changes and pass the impersonation data already embedded in the `config` parameter when it's strictly needed. The `config` parameter already has impersonation fields, so the extra `authInfo` is, apaprently, redundant.

Related kiali/kiali#3592